+++ src/syntax/abs/mod.rs
@@ -1,6 +1,7 @@
 pub use self::ast::*;
 pub use self::decl::*;
 pub use self::decl_impl::*;
+pub use self::pretty::*;
 
 /// Abstract terms.
 mod ast;
@@ -10,3 +11,4 @@ mod decl;
 mod decl_impl;
 /// Surface to abstract, scope-checking.
 pub mod desugar;
+mod pretty;